+++ lib/vgs_cg.c
@@ -166,4 +166,5 @@ void vgs_sprite(uint16_t n, BOOL visible, int16_t x, int16_t y, uint8_t size, ui
     OAM[n].rotate = 0;
     OAM[n].scale = 100;
     OAM[n].alpha = 0xFFFFFF;
+    OAM[n].ram_ptr = 0;
 }